accepts double limits now
git-svn-id: http://twcore.org/svn/trunk@753 9f8b3893-2a26-0410-9d00-a71e40fd2f77

+++ twcore/twcore/misc/lag/lagHandler.java
@@ -467,25 +467,25 @@ public class lagHandler
                     spec = true;
                     lagReport = "PING Ave. [" + averagePing + "  LIMIT: " + m_botSettings.getInt("AvePing") + "]";
                 }
-                if (m_botSettings.getInt("S2CPloss") < s2C && !spec) {
+                if (m_botSettings.getDouble("S2CPloss") < s2C && !spec) {
                     spec = true;
-                    lagReport = "PLOSS S2C. [" + s2C + "  LIMIT: " + m_botSettings.getInt("S2CPloss") + "]";
+                    lagReport = "PLOSS S2C. [" + s2C + "  LIMIT: " + m_botSettings.getDouble("S2CPloss") + "]";
                 }
-                if (m_botSettings.getInt("C2SPloss") < c2S && !spec) {
+                if (m_botSettings.getDouble("C2SPloss") < c2S && !spec) {
                     spec = true;
-                    lagReport = "PLOSS C2S. [" + c2S + "  LIMIT: " + m_botSettings.getInt("C2SPloss") + "]";
+                    lagReport = "PLOSS C2S. [" + c2S + "  LIMIT: " + m_botSettings.getDouble("C2SPloss") + "]";
                 }
-                if (m_botSettings.getInt("WeaponPloss") < s2CWeapons && !spec) {
+                if (m_botSettings.getDouble("WeaponPloss") < s2CWeapons && !spec) {
                     spec = true;
-                    lagReport = "PLOSS S2CWeapons [" + s2CWeapons + "  LIMIT: " + m_botSettings.getInt("WeaponPloss") + "]";
+                    lagReport = "PLOSS S2CWeapons [" + s2CWeapons + "  LIMIT: " + m_botSettings.getDouble("WeaponPloss") + "]";
                 }
-                if (m_botSettings.getInt("SlowS2C") < s2CSlowPercent && !spec) {
+                if (m_botSettings.getDouble("SlowS2C") < s2CSlowPercent && !spec) {
                     spec = true;
-                    lagReport = "PLOSS Slow S2C [" + s2CSlowPercent + "  LIMIT: " + m_botSettings.getInt("SlowS2C") + "]";
+                    lagReport = "PLOSS Slow S2C [" + s2CSlowPercent + "  LIMIT: " + m_botSettings.getDouble("SlowS2C") + "]";
                 }
-                if (m_botSettings.getInt("SlowC2S") < c2SSlowPercent && !spec) {
+                if (m_botSettings.getDouble("SlowC2S") < c2SSlowPercent && !spec) {
                     spec = true;
-                    lagReport = "PLOSS Slow C2S [" + c2SSlowPercent + "  LIMIT: " + m_botSettings.getInt("SlowC2S") + "]";
+                    lagReport = "PLOSS Slow C2S [" + c2SSlowPercent + "  LIMIT: " + m_botSettings.getDouble("SlowC2S") + "]";
                 }
     
                 if (tI) {
